minifac_accounts/requirements.txt
# To ensure app dependencies are ported from your virtual environment/host machine into your container, run 'pip freeze > requirements.txt' in the terminal to overwrite this file
flask==3.0.1
gunicorn==21.2.0
tabulate==0.9.0
mysql-connector-python==8.3.0
PyJWT==2.8.0